+++ packages/cli/src/config/load.js
@@ -1,3 +1,4 @@
+const path = require('path')
 const process = require('process')
 const os = require('os')
 
@@ -6,31 +7,26 @@ const { readConfig } = require('./read')
 const { validateConfig } = require('./validate')
 
 const load = async (file, required = true) => {
-  let configFile
-  let autoConfigFile = !file
-  if (file) {
-    configFile = file
-  } else {
-    configFile = await findConfig().catch(() => false)
-    if (configFile == false) {
-      if (required) {
-        throw new Error(`No configuration could be found. Please initialize a configuration via ./gallery.js run init`)
-      }
-      return {
-        configFile: file || null,
-        config: {},
-        autoConfigFile
-      }
+  const configFile = await findConfig().catch(() => false)
+  const autoConfigFile = configFile && (!file || path.resolve(file) == path.resolve(configFile))
+  if (!file && !configFile) {
+    if (required) {
+      throw new Error(`No configuration could be found. Please initialize a configuration via ./gallery.js run init`)
+    }
+    return {
+      configFile: null,
+      config: {},
+      autoConfigFile: false
     }
   }
 
   const env = {...process.env,
     HOME: process.env.HOME  || os.homedir()
   }
-  const config = await readConfig(configFile, env)
+  const config = await readConfig(file || configFile, env)
   await validateConfig(config)
   return {
-    configFile,
+    configFile: file || configFile,
     config,
     autoConfigFile,
   }