+++ packages/cli/src/utils/spawn-cli.js
@@ -1,7 +1,7 @@
 const { spawn } = require('@home-gallery/common')
 
 const Logger = require('@home-gallery/logger')
-const log = Logger('cli.task.run')
+const log = Logger('cli.spawn')
 
 const nodeBin = process.argv[0]
 const cliScript = process.argv[1]
@@ -24,16 +24,17 @@ const spawnCli = (args, options = {}, nodeArgs = []) => {
     GALLERY_LOG_LEVEL: 'trace',
     GALLERY_LOG_JSON_FORMAT: 'true'
   }
+  const t0 = Date.now()
   const cmd = `${prettyEnv(env)}${nodeBin} ${prettyArgs(runArgs)}`
   log.info(`Run cli with ${prettyArgs(args)}`)
-  const t0 = Date.now()
-  const child = spawn(nodeBin, runArgs, { ...options, env, stdio: ['pipe', 'pipe', 'inherit'] })
+  const child = spawn(nodeBin, runArgs, { ...options, env })
   log.trace(`Executing ${cmd} with pid ${child.pid}`)
+
   Logger.mergeLog(child.stdout)
 
   child.once('exit', (code, signal) => {
-    const cli = {cmd, env, nodeBin, nodeArgs, args, code}
-    log.debug({cli, duration: Date.now() - t0}, `Executed cmd ${cmd}`)
+    const spawnInfo = {env, command: nodeBin, args: runArgs, pid: child.pid, code, signal, cmd}
+    log.debug({spawn: spawnInfo, duration: Date.now() - t0}, `Executed cmd ${cmd}`)
     signal && log.info(`Cli ${prettyArgs(args)} exited by signal ${signal}`)
   })
   child.once('error', err => {
@@ -43,21 +44,6 @@ const spawnCli = (args, options = {}, nodeArgs = []) => {
   return child
 }
 
-const runCli = async (args, options = {}, nodeArgs = []) => {
-  const child = spawnCli(args, options, nodeArgs)
-
-  return new Promise((resolve, reject) => {
-    child.once('exit', (code, signal) => {
-      if (code == 0) {
-        resolve()
-      } else {
-        reject({code, signal})
-      }
-    })
-    child.once('error', err => {
-      reject(err)
-    })
-  })
+module.exports = {
+  spawnCli
 }
-
-module.exports = { runCli, spawnCli }