+++ packages/common/src/os/index.js
@@ -1,0 +1,8 @@
+const { spawn, run } = require('./spawn');
+const { ProcessManager } = require('./process-manager');
+
+module.exports = {
+  ProcessManager,
+  run,
+  spawn
+}